better naming to separate zap backend's uc api from studio rest api
